[Refactor] atomic design을 위한 폴더 정리
