+++ src/components/atoms/ContentsTitle.tsx
@@ -2,8 +2,8 @@ interface ContentsTitleProps {
   title: string;
 }
 
-const FeedContentsTitle = ({ title }: ContentsTitleProps) => {
+const ContentsTitle = ({ title }: ContentsTitleProps) => {
   return <div className='flex text-[20px] font-medium'>{title}</div>;
 };
 
-export default FeedContentsTitle;
+export default ContentsTitle;
