+++ src/components/molecules/feed/FeedBody.tsx
@@ -1,4 +1,4 @@
-import FeedContentsTitle from '@/components/molecules/contents/FeedContentsTitle';
+import ContentsTitle from '@/components/atoms/ContentsTitle';
 import ContentsBody from '@/components/molecules/feed/ContentsBody';
 import { TagItemKey, tagColors } from '@/constants/tagItem';
 import clsx from 'clsx';
@@ -12,7 +12,7 @@ interface FeedBodyProps {
 const FeedBody = ({ title, content, tags }: FeedBodyProps) => {
   return (
     <div className='flex flex-col w-full gap-[20px]'>
-      <FeedContentsTitle title={title} />
+      <ContentsTitle title={title} />
       <ContentsBody body={content} />
       <div className='flex gap-[10px]'>
         {tags?.map((tag) => (
